provide a tileset preprocessor
Provide a tileset preprocessor to separate some
CPU overhead from rendering. Using the preprocessor
causes vector tiles to be decoded separately from
rendering, which can reduce CPU overhead at the
expense of memory usage.

The effect of this optimization is increased when
using themes that process the same vector tile
elements multiple times, such as when rendering
roads with borders.