reduce creation of Paint objects
Reduce creation of paint objects to improve
rendering speed. Paint objects were created
hundreds of times when rendering a single tile.
Previously Paint objects were not cached because
they are mutable. This change introduces an
immutable model that lazily creates the paint.
A cache is introduced for Paint objects that
are mutated prior to use, such as with lines.
This cache is discarded when the tile has
been rendered.